change old api, use datasource api
