+++ server/service/dependency.go
@@ -19,14 +19,10 @@ package service
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
-	"github.com/apache/servicecomb-service-center/datasource/etcd/client"
+	"github.com/apache/servicecomb-service-center/datasource"
 	serviceUtil "github.com/apache/servicecomb-service-center/datasource/etcd/util"
 	"github.com/apache/servicecomb-service-center/pkg/log"
 	pb "github.com/apache/servicecomb-service-center/pkg/registry"
-	"github.com/apache/servicecomb-service-center/pkg/util"
-	apt "github.com/apache/servicecomb-service-center/server/core"
 	scerr "github.com/apache/servicecomb-service-center/server/scerror"
 )
 
@@ -37,7 +33,7 @@ func (s *MicroServiceService) AddDependenciesForMicroServices(ctx context.Contex
 		}, nil
 	}
 
-	resp, err := s.AddOrUpdateDependencies(ctx, in.Dependencies, false)
+	resp, err := datasource.Instance().AddOrUpdateDependencies(ctx, in.Dependencies, false)
 	return &pb.AddDependenciesResponse{
 		Response: resp,
 	}, err
@@ -50,64 +46,10 @@ func (s *MicroServiceService) CreateDependenciesForMicroServices(ctx context.Con
 		}, nil
 	}
 
-	resp, err := s.AddOrUpdateDependencies(ctx, in.Dependencies, true)
+	resp, err := datasource.Instance().AddOrUpdateDependencies(ctx, in.Dependencies, true)
 	return &pb.CreateDependenciesResponse{
 		Response: resp,
 	}, err
-}
-
-func (s *MicroServiceService) AddOrUpdateDependencies(ctx context.Context, dependencyInfos []*pb.ConsumerDependency, override bool) (*pb.Response, error) {
-	opts := make([]client.PluginOp, 0, len(dependencyInfos))
-	domainProject := util.ParseDomainProject(ctx)
-	for _, dependencyInfo := range dependencyInfos {
-		consumerFlag := util.StringJoin([]string{dependencyInfo.Consumer.Environment, dependencyInfo.Consumer.AppId, dependencyInfo.Consumer.ServiceName, dependencyInfo.Consumer.Version}, "/")
-		consumerInfo := pb.DependenciesToKeys([]*pb.MicroServiceKey{dependencyInfo.Consumer}, domainProject)[0]
-		providersInfo := pb.DependenciesToKeys(dependencyInfo.Providers, domainProject)
-
-		rsp := serviceUtil.ParamsChecker(consumerInfo, providersInfo)
-		if rsp != nil {
-			log.Errorf(nil, "put request into dependency queue failed, override: %t, consumer is %s, %s",
-				override, consumerFlag, rsp.Response.GetMessage())
-			return rsp.Response, nil
-		}
-
-		consumerID, err := serviceUtil.GetServiceID(ctx, consumerInfo)
-		if err != nil {
-			log.Errorf(err, "put request into dependency queue failed, override: %t, get consumer[%s] id failed",
-				override, consumerFlag)
-			return pb.CreateResponse(scerr.ErrInternal, err.Error()), err
-		}
-		if len(consumerID) == 0 {
-			log.Errorf(nil, "put request into dependency queue failed, override: %t, consumer[%s] does not exist",
-				override, consumerFlag)
-			return pb.CreateResponse(scerr.ErrServiceNotExists, fmt.Sprintf("Consumer %s does not exist.", consumerFlag)), nil
-		}
-
-		dependencyInfo.Override = override
-		data, err := json.Marshal(dependencyInfo)
-		if err != nil {
-			log.Errorf(err, "put request into dependency queue failed, override: %t, marshal consumer[%s] dependency failed",
-				override, consumerFlag)
-			return pb.CreateResponse(scerr.ErrInternal, err.Error()), err
-		}
-
-		id := apt.DepsQueueUUID
-		if !override {
-			id = util.GenerateUUID()
-		}
-		key := apt.GenerateConsumerDependencyQueueKey(domainProject, consumerID, id)
-		opts = append(opts, client.OpPut(client.WithStrKey(key), client.WithValue(data)))
-	}
-
-	err := client.BatchCommit(ctx, opts)
-	if err != nil {
-		log.Errorf(err, "put request into dependency queue failed, override: %t, %v", override, dependencyInfos)
-		return pb.CreateResponse(scerr.ErrInternal, err.Error()), err
-	}
-
-	log.Infof("put request into dependency queue successfully, override: %t, %v, from remote %s",
-		override, dependencyInfos, util.GetIPFromContext(ctx))
-	return pb.CreateResponse(pb.ResponseSuccess, "Create dependency successfully."), nil
 }
 
 func (s *MicroServiceService) GetProviderDependencies(ctx context.Context, in *pb.GetDependenciesRequest) (*pb.GetProDependenciesResponse, error) {
@@ -118,34 +60,8 @@ func (s *MicroServiceService) GetProviderDependencies(ctx context.Context, in *p
 			Response: pb.CreateResponse(scerr.ErrInvalidParams, err.Error()),
 		}, nil
 	}
-	domainProject := util.ParseDomainProject(ctx)
-	providerServiceID := in.ServiceId
 
-	provider, err := serviceUtil.GetService(ctx, domainProject, providerServiceID)
-	if err != nil {
-		log.Errorf(err, "GetProviderDependencies failed, provider is %s", providerServiceID)
-		return nil, err
-	}
-	if provider == nil {
-		log.Errorf(err, "GetProviderDependencies failed for provider[%s] does not exist", providerServiceID)
-		return &pb.GetProDependenciesResponse{
-			Response: pb.CreateResponse(scerr.ErrServiceNotExists, "Provider does not exist"),
-		}, nil
-	}
-
-	dr := serviceUtil.NewProviderDependencyRelation(ctx, domainProject, provider)
-	services, err := dr.GetDependencyConsumers(toDependencyFilterOptions(in)...)
-	if err != nil {
-		log.Errorf(err, "GetProviderDependencies failed, provider is %s/%s/%s/%s",
-			provider.Environment, provider.AppId, provider.ServiceName, provider.Version)
-		return &pb.GetProDependenciesResponse{
-			Response: pb.CreateResponse(scerr.ErrInternal, err.Error()),
-		}, err
-	}
-	return &pb.GetProDependenciesResponse{
-		Response:  pb.CreateResponse(pb.ResponseSuccess, "Get all consumers successful."),
-		Consumers: services,
-	}, nil
+	return datasource.Instance().SearchProviderDependency(ctx, in)
 }
 
 func (s *MicroServiceService) GetConsumerDependencies(ctx context.Context, in *pb.GetDependenciesRequest) (*pb.GetConDependenciesResponse, error) {
@@ -156,45 +72,6 @@ func (s *MicroServiceService) GetConsumerDependencies(ctx context.Context, in *p
 			Response: pb.CreateResponse(scerr.ErrInvalidParams, err.Error()),
 		}, nil
 	}
-	consumerID := in.ServiceId
-	domainProject := util.ParseDomainProject(ctx)
 
-	consumer, err := serviceUtil.GetService(ctx, domainProject, consumerID)
-	if err != nil {
-		log.Errorf(err, "GetConsumerDependencies failed, consumer is %s", consumerID)
-		return &pb.GetConDependenciesResponse{
-			Response: pb.CreateResponse(scerr.ErrInternal, err.Error()),
-		}, err
-	}
-	if consumer == nil {
-		log.Errorf(err, "GetConsumerDependencies failed for consumer[%s] does not exist", consumerID)
-		return &pb.GetConDependenciesResponse{
-			Response: pb.CreateResponse(scerr.ErrServiceNotExists, "Consumer does not exist"),
-		}, nil
-	}
-
-	dr := serviceUtil.NewConsumerDependencyRelation(ctx, domainProject, consumer)
-	services, err := dr.GetDependencyProviders(toDependencyFilterOptions(in)...)
-	if err != nil {
-		log.Errorf(err, "GetConsumerDependencies failed, consumer is %s/%s/%s/%s",
-			consumer.Environment, consumer.AppId, consumer.ServiceName, consumer.Version)
-		return &pb.GetConDependenciesResponse{
-			Response: pb.CreateResponse(scerr.ErrInternal, err.Error()),
-		}, err
-	}
-
-	return &pb.GetConDependenciesResponse{
-		Response:  pb.CreateResponse(pb.ResponseSuccess, "Get all providers successfully."),
-		Providers: services,
-	}, nil
-}
-
-func toDependencyFilterOptions(in *pb.GetDependenciesRequest) (opts []serviceUtil.DependencyRelationFilterOption) {
-	if in.SameDomain {
-		opts = append(opts, serviceUtil.WithSameDomainProject())
-	}
-	if in.NoSelf {
-		opts = append(opts, serviceUtil.WithoutSelfDependency())
-	}
-	return opts
+	return datasource.Instance().SearchConsumerDependency(ctx, in)
 }
