+++ server/service/schema.go
@@ -18,17 +18,10 @@
 package service
 
 import (
-	"fmt"
-	"strings"
-
-	"github.com/apache/servicecomb-service-center/datasource/etcd/client"
-	"github.com/apache/servicecomb-service-center/datasource/etcd/kv"
-	"github.com/apache/servicecomb-service-center/datasource/etcd/sd"
-	serviceUtil "github.com/apache/servicecomb-service-center/datasource/etcd/util"
+	"github.com/apache/servicecomb-service-center/datasource"
 	"github.com/apache/servicecomb-service-center/pkg/log"
 	pb "github.com/apache/servicecomb-service-center/pkg/registry"
 	"github.com/apache/servicecomb-service-center/pkg/util"
-	apt "github.com/apache/servicecomb-service-center/server/core"
 	"github.com/apache/servicecomb-service-center/server/plugin/quota"
 	scerr "github.com/apache/servicecomb-service-center/server/scerror"
 
@@ -44,44 +37,7 @@ func (s *MicroServiceService) GetSchemaInfo(ctx context.Context, in *pb.GetSchem
 		}, nil
 	}
 
-	domainProject := util.ParseDomainProject(ctx)
-
-	if !serviceUtil.ServiceExist(ctx, domainProject, in.ServiceId) {
-		log.Errorf(nil, "get schema[%s/%s] failed, service does not exist", in.ServiceId, in.SchemaId)
-		return &pb.GetSchemaResponse{
-			Response: pb.CreateResponse(scerr.ErrServiceNotExists, "Service does not exist."),
-		}, nil
-	}
-
-	key := apt.GenerateServiceSchemaKey(domainProject, in.ServiceId, in.SchemaId)
-	opts := append(serviceUtil.FromContext(ctx), client.WithStrKey(key))
-	resp, errDo := kv.Store().Schema().Search(ctx, opts...)
-	if errDo != nil {
-		log.Errorf(errDo, "get schema[%s/%s] failed", in.ServiceId, in.SchemaId)
-		return &pb.GetSchemaResponse{
-			Response: pb.CreateResponse(scerr.ErrUnavailableBackend, errDo.Error()),
-		}, errDo
-	}
-	if resp.Count == 0 {
-		log.Errorf(errDo, "get schema[%s/%s] failed, schema does not exists", in.ServiceId, in.SchemaId)
-		return &pb.GetSchemaResponse{
-			Response: pb.CreateResponse(scerr.ErrSchemaNotExists, "Do not have this schema info."),
-		}, nil
-	}
-
-	schemaSummary, err := getSchemaSummary(ctx, domainProject, in.ServiceId, in.SchemaId)
-	if err != nil {
-		log.Errorf(err, "get schema[%s/%s] failed, get schema summary failed", in.ServiceId, in.SchemaId)
-		return &pb.GetSchemaResponse{
-			Response: pb.CreateResponse(scerr.ErrInternal, err.Error()),
-		}, err
-	}
-
-	return &pb.GetSchemaResponse{
-		Response:      pb.CreateResponse(pb.ResponseSuccess, "Get schema info successfully."),
-		Schema:        util.BytesToStringWithNoCopy(resp.Kvs[0].Value.([]byte)),
-		SchemaSummary: schemaSummary,
-	}, nil
+	return datasource.Instance().GetSchema(ctx, in)
 }
 
 func (s *MicroServiceService) GetAllSchemaInfo(ctx context.Context, in *pb.GetAllSchemaRequest) (*pb.GetAllSchemaResponse, error) {
@@ -93,143 +49,20 @@ func (s *MicroServiceService) GetAllSchemaInfo(ctx context.Context, in *pb.GetAl
 		}, nil
 	}
 
-	domainProject := util.ParseDomainProject(ctx)
-
-	service, err := serviceUtil.GetService(ctx, domainProject, in.ServiceId)
-	if err != nil {
-		log.Errorf(err, "get service[%s] all schemas failed, get service failed", in.ServiceId)
-		return &pb.GetAllSchemaResponse{
-			Response: pb.CreateResponse(scerr.ErrInternal, err.Error()),
-		}, err
-	}
-	if service == nil {
-		log.Errorf(nil, "get service[%s] all schemas failed, service does not exist", in.ServiceId)
-		return &pb.GetAllSchemaResponse{
-			Response: pb.CreateResponse(scerr.ErrServiceNotExists, "Service does not exist."),
-		}, nil
-	}
-
-	schemasList := service.Schemas
-	if len(schemasList) == 0 {
-		return &pb.GetAllSchemaResponse{
-			Response: pb.CreateResponse(pb.ResponseSuccess, "Do not have this schema info."),
-			Schemas:  []*pb.Schema{},
-		}, nil
-	}
-
-	key := apt.GenerateServiceSchemaSummaryKey(domainProject, in.ServiceId, "")
-	opts := append(serviceUtil.FromContext(ctx), client.WithStrKey(key), client.WithPrefix())
-	resp, errDo := kv.Store().SchemaSummary().Search(ctx, opts...)
-	if errDo != nil {
-		log.Errorf(errDo, "get service[%s] all schema summaries failed", in.ServiceId)
-		return &pb.GetAllSchemaResponse{
-			Response: pb.CreateResponse(scerr.ErrUnavailableBackend, errDo.Error()),
-		}, errDo
-	}
-
-	respWithSchema := &sd.Response{}
-	if in.WithSchema {
-		key := apt.GenerateServiceSchemaKey(domainProject, in.ServiceId, "")
-		opts := append(serviceUtil.FromContext(ctx), client.WithStrKey(key), client.WithPrefix())
-		respWithSchema, errDo = kv.Store().Schema().Search(ctx, opts...)
-		if errDo != nil {
-			log.Errorf(errDo, "get service[%s] all schemas failed", in.ServiceId)
-			return &pb.GetAllSchemaResponse{
-				Response: pb.CreateResponse(scerr.ErrUnavailableBackend, errDo.Error()),
-			}, errDo
-		}
-	}
-
-	schemas := make([]*pb.Schema, 0, len(schemasList))
-	for _, schemaID := range schemasList {
-		tempSchema := &pb.Schema{}
-		tempSchema.SchemaId = schemaID
-		for _, summarySchema := range resp.Kvs {
-			_, _, schemaIDOfSummary := apt.GetInfoFromSchemaSummaryKV(summarySchema.Key)
-			if schemaID == schemaIDOfSummary {
-				tempSchema.Summary = summarySchema.Value.(string)
-			}
-		}
-
-		for _, contentSchema := range respWithSchema.Kvs {
-			_, _, schemaIDOfSchema := apt.GetInfoFromSchemaKV(contentSchema.Key)
-			if schemaID == schemaIDOfSchema {
-				tempSchema.Schema = util.BytesToStringWithNoCopy(contentSchema.Value.([]byte))
-			}
-		}
-		schemas = append(schemas, tempSchema)
-	}
-
-	return &pb.GetAllSchemaResponse{
-		Response: pb.CreateResponse(pb.ResponseSuccess, "Get all schema info successfully."),
-		Schemas:  schemas,
-	}, nil
-
+	return datasource.Instance().GetAllSchemas(ctx, in)
 }
 
 func (s *MicroServiceService) DeleteSchema(ctx context.Context, in *pb.DeleteSchemaRequest) (*pb.DeleteSchemaResponse, error) {
-	remoteIP := util.GetIPFromContext(ctx)
 	err := Validate(in)
 	if err != nil {
+		remoteIP := util.GetIPFromContext(ctx)
 		log.Errorf(err, "delete schema[%s/%s] failed, operator: %s", in.ServiceId, in.SchemaId, remoteIP)
 		return &pb.DeleteSchemaResponse{
 			Response: pb.CreateResponse(scerr.ErrInvalidParams, err.Error()),
 		}, nil
 	}
-	domainProject := util.ParseDomainProject(ctx)
 
-	if !serviceUtil.ServiceExist(ctx, domainProject, in.ServiceId) {
-		log.Errorf(nil, "delete schema[%s/%s] failed, service does not exist, operator: %s",
-			in.ServiceId, in.SchemaId, remoteIP)
-		return &pb.DeleteSchemaResponse{
-			Response: pb.CreateResponse(scerr.ErrServiceNotExists, "Service does not exist."),
-		}, nil
-	}
-
-	key := apt.GenerateServiceSchemaKey(domainProject, in.ServiceId, in.SchemaId)
-	exist, err := serviceUtil.CheckSchemaInfoExist(ctx, key)
-	if err != nil {
-		log.Errorf(err, "delete schema[%s/%s] failed, operator: %s", in.ServiceId, in.SchemaId, remoteIP)
-		return &pb.DeleteSchemaResponse{
-			Response: pb.CreateResponse(scerr.ErrInternal, err.Error()),
-		}, err
-	}
-	if !exist {
-		log.Errorf(nil, "delete schema[%s/%s] failed, schema does not exist, operator: %s",
-			in.ServiceId, in.SchemaId, remoteIP)
-		return &pb.DeleteSchemaResponse{
-			Response: pb.CreateResponse(scerr.ErrSchemaNotExists, "Schema info does not exist."),
-		}, nil
-	}
-	epSummaryKey := apt.GenerateServiceSchemaSummaryKey(domainProject, in.ServiceId, in.SchemaId)
-	opts := []client.PluginOp{
-		client.OpDel(client.WithStrKey(epSummaryKey)),
-		client.OpDel(client.WithStrKey(key)),
-	}
-
-	resp, errDo := client.Instance().TxnWithCmp(ctx, opts,
-		[]client.CompareOp{client.OpCmp(
-			client.CmpVer(util.StringToBytesWithNoCopy(apt.GenerateServiceKey(domainProject, in.ServiceId))),
-			client.CmpNotEqual, 0)},
-		nil)
-	if errDo != nil {
-		log.Errorf(errDo, "delete schema[%s/%s] failed, operator: %s", in.ServiceId, in.SchemaId, remoteIP)
-		return &pb.DeleteSchemaResponse{
-			Response: pb.CreateResponse(scerr.ErrUnavailableBackend, errDo.Error()),
-		}, errDo
-	}
-	if !resp.Succeeded {
-		log.Errorf(nil, "delete schema[%s/%s] failed, service does not exist, operator: %s",
-			in.ServiceId, in.SchemaId, remoteIP)
-		return &pb.DeleteSchemaResponse{
-			Response: pb.CreateResponse(scerr.ErrServiceNotExists, "Service does not exist."),
-		}, nil
-	}
-
-	log.Infof("delete schema[%s/%s] info successfully, operator: %s", in.ServiceId, in.SchemaId, remoteIP)
-	return &pb.DeleteSchemaResponse{
-		Response: pb.CreateResponse(pb.ResponseSuccess, "Delete schema info successfully."),
-	}, nil
+	return datasource.Instance().DeleteSchema(ctx, in)
 }
 
 // ModifySchemas covers all the schemas of a service.
@@ -245,267 +78,15 @@ func (s *MicroServiceService) DeleteSchema(ctx context.Context, in *pb.DeleteSch
 // the new schemaID will be automatically added to the service information.
 // Schema is allowed to add/delete/modify.
 func (s *MicroServiceService) ModifySchemas(ctx context.Context, in *pb.ModifySchemasRequest) (*pb.ModifySchemasResponse, error) {
-	remoteIP := util.GetIPFromContext(ctx)
 	err := Validate(in)
 	if err != nil {
+		remoteIP := util.GetIPFromContext(ctx)
 		log.Errorf(err, "modify service[%s] schemas failed, operator: %s", in.ServiceId, remoteIP)
 		return &pb.ModifySchemasResponse{
 			Response: pb.CreateResponse(scerr.ErrInvalidParams, "Invalid request."),
 		}, nil
 	}
-	serviceID := in.ServiceId
-
-	domainProject := util.ParseDomainProject(ctx)
-
-	service, err := serviceUtil.GetService(ctx, domainProject, serviceID)
-	if err != nil {
-		log.Errorf(err, "modify service[%s] schemas failed, get service failed, operator: %s", serviceID, remoteIP)
-		return &pb.ModifySchemasResponse{
-			Response: pb.CreateResponse(scerr.ErrInternal, err.Error()),
-		}, err
-	}
-	if service == nil {
-		log.Errorf(nil, "modify service[%s] schemas failed, service does not exist, operator: %s",
-			serviceID, remoteIP)
-		return &pb.ModifySchemasResponse{
-			Response: pb.CreateResponse(scerr.ErrServiceNotExists, "Service does not exist."),
-		}, nil
-	}
-
-	respErr := s.modifySchemas(ctx, domainProject, service, in.Schemas)
-	if respErr != nil {
-		log.Errorf(nil, "modify service[%s] schemas failed, operator: %s", serviceID, remoteIP)
-		resp := &pb.ModifySchemasResponse{
-			Response: pb.CreateResponseWithSCErr(respErr),
-		}
-		if respErr.InternalError() {
-			return resp, respErr
-		}
-		return resp, nil
-	}
-
-	return &pb.ModifySchemasResponse{
-		Response: pb.CreateResponse(pb.ResponseSuccess, "modify schemas info successfully."),
-	}, nil
-}
-
-func schemasAnalysis(schemas []*pb.Schema, schemasFromDb []*pb.Schema, schemaIDsInService []string) (
-	[]*pb.Schema, []*pb.Schema, []*pb.Schema, []string) {
-	needUpdateSchemas := make([]*pb.Schema, 0, len(schemas))
-	needAddSchemas := make([]*pb.Schema, 0, len(schemas))
-	needDeleteSchemas := make([]*pb.Schema, 0, len(schemasFromDb))
-	nonExistSchemaIds := make([]string, 0, len(schemas))
-
-	duplicate := make(map[string]struct{})
-	for _, schema := range schemas {
-		if _, ok := duplicate[schema.SchemaId]; ok {
-			continue
-		}
-		duplicate[schema.SchemaId] = struct{}{}
-
-		exist := false
-		for _, schemaFromDb := range schemasFromDb {
-			if schema.SchemaId == schemaFromDb.SchemaId {
-				needUpdateSchemas = append(needUpdateSchemas, schema)
-				exist = true
-				break
-			}
-		}
-		if !exist {
-			needAddSchemas = append(needAddSchemas, schema)
-		}
-
-		exist = false
-		for _, schemaID := range schemaIDsInService {
-			if schema.SchemaId == schemaID {
-				exist = true
-			}
-		}
-		if !exist {
-			nonExistSchemaIds = append(nonExistSchemaIds, schema.SchemaId)
-		}
-	}
-
-	for _, schemaFromDb := range schemasFromDb {
-		exist := false
-		for _, schema := range schemas {
-			if schema.SchemaId == schemaFromDb.SchemaId {
-				exist = true
-				break
-			}
-		}
-		if !exist {
-			needDeleteSchemas = append(needDeleteSchemas, schemaFromDb)
-		}
-	}
-
-	return needUpdateSchemas, needAddSchemas, needDeleteSchemas, nonExistSchemaIds
-}
-
-func (s *MicroServiceService) modifySchemas(ctx context.Context, domainProject string, service *pb.MicroService, schemas []*pb.Schema) *scerr.Error {
-	remoteIP := util.GetIPFromContext(ctx)
-	serviceID := service.ServiceId
-	schemasFromDatabase, err := GetSchemasFromDatabase(ctx, domainProject, serviceID)
-	if err != nil {
-		log.Errorf(nil, "modify service[%s] schemas failed, get schemas failed, operator: %s",
-			serviceID, remoteIP)
-		return scerr.NewError(scerr.ErrUnavailableBackend, err.Error())
-	}
-
-	needUpdateSchemas, needAddSchemas, needDeleteSchemas, nonExistSchemaIds := schemasAnalysis(schemas, schemasFromDatabase, service.Schemas)
-
-	pluginOps := make([]client.PluginOp, 0)
-	if !s.isSchemaEditable(service) {
-		if len(service.Schemas) == 0 {
-			res := quota.NewApplyQuotaResource(quota.SchemaQuotaType, domainProject, serviceID, int64(len(nonExistSchemaIds)))
-			rst := quota.Apply(ctx, res)
-			errQuota := rst.Err
-			if errQuota != nil {
-				log.Errorf(errQuota, "modify service[%s] schemas failed, operator: %s", serviceID, remoteIP)
-				return errQuota
-			}
-
-			service.Schemas = nonExistSchemaIds
-			opt, err := serviceUtil.UpdateService(domainProject, serviceID, service)
-			if err != nil {
-				log.Errorf(err, "modify service[%s] schemas failed, update service.Schemas failed, operator: %s",
-					serviceID, remoteIP)
-				return scerr.NewError(scerr.ErrInternal, err.Error())
-			}
-			pluginOps = append(pluginOps, opt)
-		} else {
-			if len(nonExistSchemaIds) != 0 {
-				errInfo := fmt.Errorf("Non-existent schemaIDs %v", nonExistSchemaIds)
-				log.Errorf(errInfo, "modify service[%s] schemas failed, operator: %s", serviceID, remoteIP)
-				return scerr.NewError(scerr.ErrUndefinedSchemaID, errInfo.Error())
-			}
-			for _, needUpdateSchema := range needUpdateSchemas {
-				exist, err := isExistSchemaSummary(ctx, domainProject, serviceID, needUpdateSchema.SchemaId)
-				if err != nil {
-					return scerr.NewError(scerr.ErrInternal, err.Error())
-				}
-				if !exist {
-					opts := schemaWithDatabaseOpera(client.OpPut, domainProject, serviceID, needUpdateSchema)
-					pluginOps = append(pluginOps, opts...)
-				} else {
-					log.Warnf("schema[%s/%s] and it's summary already exist, skip to update, operator: %s",
-						serviceID, needUpdateSchema.SchemaId, remoteIP)
-				}
-			}
-		}
-
-		for _, schema := range needAddSchemas {
-			log.Infof("add new schema[%s/%s], operator: %s", serviceID, schema.SchemaId, remoteIP)
-			opts := schemaWithDatabaseOpera(client.OpPut, domainProject, service.ServiceId, schema)
-			pluginOps = append(pluginOps, opts...)
-		}
-	} else {
-		quotaSize := len(needAddSchemas) - len(needDeleteSchemas)
-		if quotaSize > 0 {
-			res := quota.NewApplyQuotaResource(quota.SchemaQuotaType, domainProject, serviceID, int64(quotaSize))
-			rst := quota.Apply(ctx, res)
-			err := rst.Err
-			if err != nil {
-				log.Errorf(err, "modify service[%s] schemas failed, operator: %s", serviceID, remoteIP)
-				return err
-			}
-		}
-
-		var schemaIDs []string
-		for _, schema := range needAddSchemas {
-			log.Infof("add new schema[%s/%s], operator: %s", serviceID, schema.SchemaId, remoteIP)
-			opts := schemaWithDatabaseOpera(client.OpPut, domainProject, service.ServiceId, schema)
-			pluginOps = append(pluginOps, opts...)
-			schemaIDs = append(schemaIDs, schema.SchemaId)
-		}
-
-		for _, schema := range needUpdateSchemas {
-			log.Infof("update schema[%s/%s], operator: %s", serviceID, schema.SchemaId, remoteIP)
-			opts := schemaWithDatabaseOpera(client.OpPut, domainProject, serviceID, schema)
-			pluginOps = append(pluginOps, opts...)
-			schemaIDs = append(schemaIDs, schema.SchemaId)
-		}
-
-		for _, schema := range needDeleteSchemas {
-			log.Infof("delete non-existent schema[%s/%s], operator: %s", serviceID, schema.SchemaId, remoteIP)
-			opts := schemaWithDatabaseOpera(client.OpDel, domainProject, serviceID, schema)
-			pluginOps = append(pluginOps, opts...)
-		}
-
-		service.Schemas = schemaIDs
-		opt, err := serviceUtil.UpdateService(domainProject, serviceID, service)
-		if err != nil {
-			log.Errorf(err, "modify service[%s] schemas failed, update service.Schemas failed, operator: %s",
-				serviceID, remoteIP)
-			return scerr.NewError(scerr.ErrInternal, err.Error())
-		}
-		pluginOps = append(pluginOps, opt)
-	}
-
-	if len(pluginOps) != 0 {
-		resp, err := client.BatchCommitWithCmp(ctx, pluginOps,
-			[]client.CompareOp{client.OpCmp(
-				client.CmpVer(util.StringToBytesWithNoCopy(apt.GenerateServiceKey(domainProject, serviceID))),
-				client.CmpNotEqual, 0)},
-			nil)
-		if err != nil {
-			return scerr.NewError(scerr.ErrUnavailableBackend, err.Error())
-		}
-		if !resp.Succeeded {
-			return scerr.NewError(scerr.ErrServiceNotExists, "Service does not exist.")
-		}
-	}
-	return nil
-}
-
-func (s *MicroServiceService) isSchemaEditable(service *pb.MicroService) bool {
-	return (len(service.Environment) != 0 && service.Environment != pb.ENV_PROD) || s.schemaEditable
-}
-
-func isExistSchemaID(service *pb.MicroService, schemas []*pb.Schema) bool {
-	serviceSchemaIds := service.Schemas
-	for _, schema := range schemas {
-		if !containsValueInSlice(serviceSchemaIds, schema.SchemaId) {
-			log.Errorf(nil, "schema[%s/%s] does not exist schemaID", service.ServiceId, schema.SchemaId)
-			return false
-		}
-	}
-	return true
-}
-
-func schemaWithDatabaseOpera(invoke client.Operation, domainProject string, serviceID string, schema *pb.Schema) []client.PluginOp {
-	pluginOps := make([]client.PluginOp, 0)
-	key := apt.GenerateServiceSchemaKey(domainProject, serviceID, schema.SchemaId)
-	opt := invoke(client.WithStrKey(key), client.WithStrValue(schema.Schema))
-	pluginOps = append(pluginOps, opt)
-	keySummary := apt.GenerateServiceSchemaSummaryKey(domainProject, serviceID, schema.SchemaId)
-	opt = invoke(client.WithStrKey(keySummary), client.WithStrValue(schema.Summary))
-	pluginOps = append(pluginOps, opt)
-	return pluginOps
-}
-
-func GetSchemasFromDatabase(ctx context.Context, domainProject string, serviceID string) ([]*pb.Schema, error) {
-	key := apt.GenerateServiceSchemaKey(domainProject, serviceID, "")
-	resp, err := kv.Store().Schema().Search(ctx,
-		client.WithPrefix(),
-		client.WithStrKey(key))
-	if err != nil {
-		log.Errorf(err, "get service[%s]'s schema failed", serviceID)
-		return nil, err
-	}
-	schemas := make([]*pb.Schema, 0, len(resp.Kvs))
-	for _, kv := range resp.Kvs {
-		key := util.BytesToStringWithNoCopy(kv.Key)
-		tmp := strings.Split(key, "/")
-		schemaID := tmp[len(tmp)-1]
-		schema := util.BytesToStringWithNoCopy(kv.Value.([]byte))
-		schemaStruct := &pb.Schema{
-			SchemaId: schemaID,
-			Schema:   schema,
-		}
-		schemas = append(schemas, schemaStruct)
-	}
-	return schemas, nil
+	return datasource.Instance().ModifySchemas(ctx, in)
 }
 
 // ModifySchema modifies a specific schema
@@ -519,7 +100,6 @@ func GetSchemasFromDatabase(ctx context.Context, domainProject string, serviceID
 // the new schemaID will be automatically added to the service information.
 // Schema is allowed to add/modify.
 func (s *MicroServiceService) ModifySchema(ctx context.Context, request *pb.ModifySchemaRequest) (*pb.ModifySchemaResponse, error) {
-	remoteIP := util.GetIPFromContext(ctx)
 	domainProject := util.ParseDomainProject(ctx)
 	respErr := s.canModifySchema(ctx, domainProject, request)
 	if respErr != nil {
@@ -532,30 +112,7 @@ func (s *MicroServiceService) ModifySchema(ctx context.Context, request *pb.Modi
 		return resp, nil
 	}
 
-	serviceID := request.ServiceId
-	schemaID := request.SchemaId
-
-	schema := pb.Schema{
-		SchemaId: schemaID,
-		Summary:  request.Summary,
-		Schema:   request.Schema,
-	}
-	err := s.modifySchema(ctx, serviceID, &schema)
-	if err != nil {
-		log.Errorf(err, "modify schema[%s/%s] failed, operator: %s", serviceID, schemaID, remoteIP)
-		resp := &pb.ModifySchemaResponse{
-			Response: pb.CreateResponseWithSCErr(err),
-		}
-		if err.InternalError() {
-			return resp, err
-		}
-		return resp, nil
-	}
-
-	log.Infof("modify schema[%s/%s] successfully, operator: %s", serviceID, schemaID, remoteIP)
-	return &pb.ModifySchemaResponse{
-		Response: pb.CreateResponse(pb.ResponseSuccess, "modify schema info success"),
-	}, nil
+	return datasource.Instance().ModifySchema(ctx, request)
 }
 
 func (s *MicroServiceService) canModifySchema(ctx context.Context, domainProject string, in *pb.ModifySchemaRequest) *scerr.Error {
@@ -584,145 +141,4 @@ func (s *MicroServiceService) canModifySchema(ctx context.Context, domainProject
 		log.Warnf("schema[%s/%s]'s summary is empty, operator: %s", serviceID, schemaID, remoteIP)
 	}
 	return nil
-}
-
-func (s *MicroServiceService) modifySchema(ctx context.Context, serviceID string, schema *pb.Schema) *scerr.Error {
-	remoteIP := util.GetIPFromContext(ctx)
-	domainProject := util.ParseDomainProject(ctx)
-	schemaID := schema.SchemaId
-
-	service, err := serviceUtil.GetService(ctx, domainProject, serviceID)
-	if err != nil {
-		log.Errorf(err, "modify schema[%s/%s] failed, get service failed, operator: %s",
-			serviceID, schemaID, remoteIP)
-		return scerr.NewError(scerr.ErrInternal, err.Error())
-	}
-	if service == nil {
-		log.Errorf(nil, "modify schema[%s/%s] failed, service does not exist, operator: %s",
-			serviceID, schemaID, remoteIP)
-		return scerr.NewError(scerr.ErrServiceNotExists, "Service does not exist")
-	}
-
-	var pluginOps []client.PluginOp
-	isExist := isExistSchemaID(service, []*pb.Schema{schema})
-
-	if !s.isSchemaEditable(service) {
-		if len(service.Schemas) != 0 && !isExist {
-			return scerr.NewError(scerr.ErrUndefinedSchemaID, "Non-existent schemaID can't be added in "+pb.ENV_PROD)
-		}
-
-		key := apt.GenerateServiceSchemaKey(domainProject, serviceID, schemaID)
-		respSchema, err := kv.Store().Schema().Search(ctx, client.WithStrKey(key), client.WithCountOnly())
-		if err != nil {
-			log.Errorf(err, "modify schema[%s/%s] failed, get schema summary failed, operator: %s",
-				serviceID, schemaID, remoteIP)
-			return scerr.NewError(scerr.ErrUnavailableBackend, err.Error())
-		}
-
-		if respSchema.Count != 0 {
-			if len(schema.Summary) == 0 {
-				log.Errorf(err, "%s mode, schema[%s/%s] already exists, can not be changed, operator: %s",
-					pb.ENV_PROD, serviceID, schemaID, remoteIP)
-				return scerr.NewError(scerr.ErrModifySchemaNotAllow, "schema already exist, can not be changed in "+pb.ENV_PROD)
-			}
-
-			exist, err := isExistSchemaSummary(ctx, domainProject, serviceID, schemaID)
-			if err != nil {
-				log.Errorf(err, "check schema[%s/%s] summary existence failed, operator: %s",
-					serviceID, schemaID, remoteIP)
-				return scerr.NewError(scerr.ErrInternal, err.Error())
-			}
-			if exist {
-				log.Errorf(err, "%s mode, schema[%s/%s] already exist, can not be changed, operator: %s",
-					pb.ENV_PROD, serviceID, schemaID, remoteIP)
-				return scerr.NewError(scerr.ErrModifySchemaNotAllow, "schema already exist, can not be changed in "+pb.ENV_PROD)
-			}
-		}
-
-		if len(service.Schemas) == 0 {
-			service.Schemas = append(service.Schemas, schemaID)
-			opt, err := serviceUtil.UpdateService(domainProject, serviceID, service)
-			if err != nil {
-				log.Errorf(err, "modify schema[%s/%s] failed, update service.Schemas failed, operator: %s",
-					serviceID, schemaID, remoteIP)
-				return scerr.NewError(scerr.ErrInternal, err.Error())
-			}
-			pluginOps = append(pluginOps, opt)
-		}
-	} else {
-		if !isExist {
-			service.Schemas = append(service.Schemas, schemaID)
-			opt, err := serviceUtil.UpdateService(domainProject, serviceID, service)
-			if err != nil {
-				log.Errorf(err, "modify schema[%s/%s] failed, update service.Schemas failed, operator: %s",
-					serviceID, schemaID, remoteIP)
-				return scerr.NewError(scerr.ErrInternal, err.Error())
-			}
-			pluginOps = append(pluginOps, opt)
-		}
-	}
-
-	opts := CommitSchemaInfo(domainProject, serviceID, schema)
-	pluginOps = append(pluginOps, opts...)
-
-	resp, err := client.Instance().TxnWithCmp(ctx, pluginOps,
-		[]client.CompareOp{client.OpCmp(
-			client.CmpVer(util.StringToBytesWithNoCopy(apt.GenerateServiceKey(domainProject, serviceID))),
-			client.CmpNotEqual, 0)},
-		nil)
-	if err != nil {
-		return scerr.NewError(scerr.ErrUnavailableBackend, err.Error())
-	}
-	if !resp.Succeeded {
-		return scerr.NewError(scerr.ErrServiceNotExists, "Service does not exist.")
-	}
-	return nil
-}
-
-func isExistSchemaSummary(ctx context.Context, domainProject, serviceID, schemaID string) (bool, error) {
-	key := apt.GenerateServiceSchemaSummaryKey(domainProject, serviceID, schemaID)
-	resp, err := kv.Store().SchemaSummary().Search(ctx, client.WithStrKey(key), client.WithCountOnly())
-	if err != nil {
-		return true, err
-	}
-	if resp.Count == 0 {
-		return false, nil
-	}
-	return true, nil
-}
-
-func CommitSchemaInfo(domainProject string, serviceID string, schema *pb.Schema) []client.PluginOp {
-	if len(schema.Summary) != 0 {
-		return schemaWithDatabaseOpera(client.OpPut, domainProject, serviceID, schema)
-	}
-	key := apt.GenerateServiceSchemaKey(domainProject, serviceID, schema.SchemaId)
-	opt := client.OpPut(client.WithStrKey(key), client.WithStrValue(schema.Schema))
-	return []client.PluginOp{opt}
-}
-
-func containsValueInSlice(in []string, value string) bool {
-	if in == nil || len(value) == 0 {
-		return false
-	}
-	for _, i := range in {
-		if i == value {
-			return true
-		}
-	}
-	return false
-}
-
-func getSchemaSummary(ctx context.Context, domainProject string, serviceID string, schemaID string) (string, error) {
-	key := apt.GenerateServiceSchemaSummaryKey(domainProject, serviceID, schemaID)
-	resp, err := kv.Store().SchemaSummary().Search(ctx,
-		client.WithStrKey(key),
-	)
-	if err != nil {
-		log.Errorf(err, "get schema[%s/%s] summary failed", serviceID, schemaID)
-		return "", err
-	}
-	if len(resp.Kvs) == 0 {
-		return "", nil
-	}
-	return resp.Kvs[0].Value.(string), nil
 }
